rename _namespace
