+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/_namespace.py
